Refactor to match upstream support tracks better
Instead of tracking individual minor releases, this tracks "stable", "rc", and "1.10" (hopefully someday renamed to "lts").

This matches https://julialang.org/downloads/manual-downloads/ and https://docs.julialang.org/en/v1/manual/faq/#Do-I-want-to-use-the-Stable,-LTS,-or-nightly-version-of-Julia?

This then reflects that 1.11 is EOL (and has been since 1.12 went GA), and does so automatically (when 1.13 is GA, "stable" will auto-update to it, and "rc" will drop off until 1.14.0-alpha1+).